update remotestorage to 0.12.1
